Add test for Pie and Legend animation
